v1.1.0 - Seperation of Managers and Classes
