+++ src/main/java/dk/dtu/cdio2/Game.java
@@ -1,21 +1,43 @@
 package dk.dtu.cdio2;
 
 import dk.dtu.cdio2.actions.*;
+import dk.dtu.cdio2.managers.*;
 
 /**
  * Main class for the program
  */
 public class Game {
     private final static LanguageManager lm = new LanguageManager();
-    private final static GUIManager gm = new GUIManager(lm);
-    private final static PlayerManager pm = new PlayerManager(gm);
+    private final static GUIManager gm = new GUIManager();
+    private final static PlayerManager pm = new PlayerManager();
     private final static DiceManager dm = new DiceManager();
-    private final static ActionManager am = new ActionManager(dm, gm, pm);
+    private final static ActionManager actm = new ActionManager();
+    private final static AccountManager accm = new AccountManager();
 
     private final static int startingBalance = 1000;
     private final static int finalBalance = 3000;
     private static boolean isPlaying = true;
     public final static boolean debug = ((System.getenv("debug") != null) || (System.getProperty("debug") != null));
+
+    // Getter functions for the managers
+    public static LanguageManager getLanguageManager() {
+        return lm;
+    }
+    public static GUIManager getGUIManager() {
+        return gm;
+    }
+    public static PlayerManager getPlayerManager() {
+        return pm;
+    }
+    public static DiceManager getDiceManager() {
+        return dm;
+    }
+    public static ActionManager getActionManager() {
+        return actm;
+    }
+    public static AccountManager getAccountManager() {
+        return accm;
+    }
 
     public static void main(String[] args) {
         // Choose a language
@@ -30,7 +52,7 @@ public class Game {
         // Create the players
         for (int i = 1; i <= 2; i++) {
             String playerName = gm.getUserString(lm.getString("enter_player_name").replace("{player_number}", Integer.toString(i)));
-            PlayerManager.Player player = pm.createPlayer(playerName, startingBalance);
+            Player player = pm.createPlayer(playerName, startingBalance);
             gm.createGUIPlayer(player.getID(), playerName, player.getMoney());
         }
 
@@ -45,13 +67,13 @@ public class Game {
     }
 
     private static void play() {
-        Action rollAction = am.getAction(1);
+        Action rollAction = Game.getActionManager().getAction(1);
 
         int playerWon = 0;
         while (playerWon == 0) {
             // Loop through players
             for (int playerID : pm.getPlayerIDs()) {
-                PlayerManager.Player player = pm.getPlayer(playerID);
+                Player player = pm.getPlayer(playerID);
 
                 rollAction.doAction(playerID);
