+++ src/main/java/dk/dtu/cdio2/actions/RollAction.java
@@ -1,38 +1,31 @@
 package dk.dtu.cdio2.actions;
 
-import dk.dtu.cdio2.ActionManager;
-import dk.dtu.cdio2.DiceManager.DiceCup;
-import dk.dtu.cdio2.GUIManager;
+import dk.dtu.cdio2.DiceCup;
+import dk.dtu.cdio2.fields.Field;
+import dk.dtu.cdio2.Player;
 import dk.dtu.cdio2.Game;
-import dk.dtu.cdio2.PlayerManager;
 
 public class RollAction extends Action {
     DiceCup dc;
-    GUIManager gm;
-    PlayerManager pm;
-    ActionManager am;
 
-    public RollAction(DiceCup dc, GUIManager gm, PlayerManager pm, ActionManager am) {
+    public RollAction() {
         super();
-        this.dc = dc;
-        this.gm = gm;
-        this.pm = pm;
-        this.am = am;
+        this.dc = Game.getDiceManager().createDiceCup();
     }
 
     @Override
     public void doAction(Integer playerID) {
-        PlayerManager.Player player = pm.getPlayer(playerID);
+        Player player = Game.getPlayerManager().getPlayer(playerID);
 
         if (!Game.debug) {
-            gm.waitUserRoll(player.getName());
+            Game.getGUIManager().waitUserRoll(player.getName());
         }
         dc.raffleCup();
 
         int[] diceValues = dc.getDiceValues();
-        gm.updateDice(diceValues[0], diceValues[1]);
+        Game.getGUIManager().updateDice(diceValues[0], diceValues[1]);
 
-        GUIManager.Field field = gm.movePlayerField(player.getID(), dc.getSum());
-        field.doLandingAction(pm, am, player.getID());
+        Field field = Game.getGUIManager().movePlayerField(player.getID(), dc.getSum());
+        field.doLandingAction(player.getID());
     }
 }
