+++ src/main/java/dk/dtu/cdio2/managers/ActionManager.java
@@ -1,12 +1,12 @@
-package dk.dtu.cdio2;
+package dk.dtu.cdio2.managers;
 
 import dk.dtu.cdio2.actions.*;
 
 public class ActionManager {
     private Action[] actions;
 
-    public ActionManager(DiceManager dm, GUIManager gm, PlayerManager pm) {
-        actions = new Action[]{ new RollAction(dm.createDiceCup(), gm, pm, this ) };
+    public ActionManager() {
+        actions = new Action[]{ new RollAction() };
     }
 
     public Action getAction(int actionID) {
