+++ src/test/java/dk/dtu/cdio2/TestDiceManager.java
@@ -1,5 +1,6 @@
 package dk.dtu.cdio2;
 
+import dk.dtu.cdio2.managers.DiceManager;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 import java.util.HashMap;
@@ -17,7 +18,7 @@ public class TestDiceManager {
      */
     @Test
     public void testDiceCupAndDieCreation() {
-        DiceManager.DiceCup testCup = dm.createDiceCup();
+        DiceCup testCup = dm.createDiceCup();
         assertArrayEquals(new int[] {6, 6}, testCup.getDiceValues());
     }
 
@@ -27,7 +28,7 @@ public class TestDiceManager {
      */
     @Test
     public void testDiceCupRaffle() {
-        DiceManager.DiceCup testCup = dm.createDiceCup();
+        DiceCup testCup = dm.createDiceCup();
 
         // We test the dice cup for 1000 throws
         for (int i = 0; i < 1000; i++) {
@@ -59,7 +60,7 @@ public class TestDiceManager {
         // We also keep an integer to keep track of how many equal throws are made
         int equalThrows = 0;
 
-        DiceManager.DiceCup testCup = dm.createDiceCup();
+        DiceCup testCup = dm.createDiceCup();
 
         // We test the dice cup for x amount of throws
         int throwAmount = 10000000;
