+++ src/test/java/dk/dtu/cdio2/TestLanguageManager.java
@@ -1,5 +1,6 @@
 package dk.dtu.cdio2;
 
+import dk.dtu.cdio2.managers.LanguageManager;
 import org.junit.jupiter.api.Test;
 
 import java.util.Locale;
